Version 2.1.0: Renamed built files to for better clarity/ branding. Changes to test files to reflect the new built files.
